refactor(entities): weatherエンティティの公開APIを作成し境界を明確化
weatherエンティティのバレルファイル(index.ts)を作成し、ワイルドカードを使用せずに名前付きエクスポートを定義した。
featuresおよびwidgets層からのインポートを、内部ファイルの直接参照から公開API（@/entities/weather）経由に変更することで、FSDのレイヤーカプセル化を厳格化した。
